+++ pubnub.go
@@ -12,7 +12,7 @@ import (
 // Default constants
 const (
 	// Version :the version of the SDK
-	Version = "4.9.1"
+	Version = "4.10.0"
 	// MaxSequence for publish messages
 	MaxSequence = 65535
 )